backend/requirements.txt
# API & Server
fastapi==0.136.1
uvicorn==0.30.6
python-multipart==0.0.28

# Frontend
streamlit==1.57.0

# Database
sqlalchemy==2.0.35

# Data Science
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
plotly==5.24.1

# AI / LLM
langchain==0.3.1
langchain-groq==0.2.0
langchain-community==0.3.1

# RAG / Vector Store
faiss-cpu==1.8.0
sentence-transformers==3.1.1

# Report Generation
fpdf2==2.7.9

# Environment
python-dotenv==1.0.1

# Forecasting
statsmodels==0.14.4